ensure request count read only once
