+++ controllers/experiment_controller_test.go
@@ -169,231 +169,6 @@ var _ = Describe("Experiment Validation", func() {
 	})
 })
 
-var _ = Describe("Metrics", func() {
-	var jqe string = "expr"
-	var url string = "url"
-
-	var testName string
-	var testNamespace, metricsNamespace string
-	var goodObjective, goodObjective2, badObjective, reward *v2alpha2.Metric
-	BeforeEach(func() {
-		testNamespace = "default"
-		metricsNamespace = "metric-namespace"
-
-		k8sClient.DeleteAllOf(ctx(), &v2alpha2.Experiment{}, client.InNamespace(testNamespace))
-		k8sClient.DeleteAllOf(ctx(), &v2alpha2.Metric{}, client.InNamespace(testNamespace))
-		k8sClient.DeleteAllOf(ctx(), &v2alpha2.Metric{}, client.InNamespace(metricsNamespace))
-
-		By("Providing a request-count metric")
-		m := v2alpha2.NewMetric("request-count", metricsNamespace).
-			WithType(v2alpha2.CounterMetricType).
-			WithParams([]v2alpha2.NamedValue{{
-				Name:  "param",
-				Value: "value",
-			}}).
-			WithProvider("prometheus").
-			WithJQExpression(&jqe).
-			WithURLTemplate(&url).
-			Build()
-		Expect(k8sClient.Create(ctx(), m)).Should(Succeed())
-		goodObjective2 = v2alpha2.NewMetric("objective-with-good-reference-2", metricsNamespace).
-			WithType(v2alpha2.CounterMetricType).
-			WithParams([]v2alpha2.NamedValue{{
-				Name:  "param",
-				Value: "value",
-			}}).
-			WithProvider("prometheus").
-			WithJQExpression(&jqe).
-			WithURLTemplate(&url).
-			WithSampleSize("request-count").
-			Build()
-		Expect(k8sClient.Create(ctx(), goodObjective2)).Should(Succeed())
-		By("creating an objective that does not reference the request-count")
-		goodObjective = v2alpha2.NewMetric("objective-with-good-reference", "default").
-			WithType(v2alpha2.CounterMetricType).
-			WithParams([]v2alpha2.NamedValue{{
-				Name:  "param",
-				Value: "value",
-			}}).
-			WithProvider("prometheus").
-			WithJQExpression(&jqe).
-			WithURLTemplate(&url).
-			WithSampleSize(metricsNamespace + "/request-count").
-			Build()
-		Expect(k8sClient.Create(ctx(), goodObjective)).Should(Succeed())
-		By("creating an objective that references request-count")
-		badObjective = v2alpha2.NewMetric("objective-with-bad-reference", "default").
-			WithType(v2alpha2.CounterMetricType).
-			WithParams([]v2alpha2.NamedValue{{
-				Name:  "param",
-				Value: "value",
-			}}).
-			WithProvider("prometheus").
-			WithJQExpression(&jqe).
-			WithURLTemplate(&url).
-			WithSampleSize("request-count").
-			Build()
-		Expect(k8sClient.Create(ctx(), badObjective)).Should(Succeed())
-		reward = v2alpha2.NewMetric("rwrd", "default").
-			WithType(v2alpha2.CounterMetricType).
-			WithParams([]v2alpha2.NamedValue{{
-				Name:  "param",
-				Value: "value",
-			}}).
-			WithProvider("prometheus").
-			WithJQExpression(&jqe).
-			WithURLTemplate(&url).
-			Build()
-		Expect(k8sClient.Create(ctx(), reward)).Should(Succeed())
-	})
-
-	Context("When creating an experiment referencing valid metrics", func() {
-		// experiment (in default namespace) refers to metric "objective-with-good-reference"
-		// which has a sampleSize "metricNamespace/request-count" which is correct
-		It("Should successfully read the metrics and proceed", func() {
-			By("Creating experiment")
-			testName = "valid-reference"
-			experiment := v2alpha2.NewExperiment(testName, testNamespace).
-				WithTarget("target").
-				WithTestingPattern(v2alpha2.TestingPatternCanary).
-				WithRequestCount(metricsNamespace+"/request-count").
-				WithObjective(*goodObjective, nil, nil, false).
-				WithReward(*reward, v2alpha2.PreferredDirectionHigher).
-				Build()
-			Expect(k8sClient.Create(ctx(), experiment)).Should(Succeed())
-			By("Checking that it starts Running")
-			// this assumes that it runs for a while
-			Eventually(func() bool {
-				return containsSubString(events, "Advanced to Running") //v2alpha2.ReasonStageAdvanced)
-			}, 5).Should(BeTrue())
-		})
-	})
-
-	Context("failed start handler", func() {
-		Specify("experiment teminated in a failed state", func() {
-			By("Creating an experiment with a start handler")
-			name, target := "has-failing-handler", "has-failing-handler"
-			iterations, loops := int32(2), int32(1)
-			handler := "start"
-			experiment := v2alpha2.NewExperiment(name, testNamespace).
-				WithTarget(target).
-				WithTestingPattern(v2alpha2.TestingPatternConformance).
-				WithAction(handler, []v2alpha2.TaskSpec{}).
-				WithRequestCount(metricsNamespace+"/request-count").
-				WithDuration(30, iterations, loops).
-				WithBaselineVersion("baseline", nil).
-				Build()
-
-			Expect(k8sClient.Create(ctx(), experiment)).Should(Succeed())
-			Eventually(func() bool { return fails(name, testNamespace) }, 5).Should(BeTrue())
-		})
-	})
-
-	Context("When creating an experiment which refers to a non-existing metric", func() {
-		// experiment (in default ns) refers to metric "request-count" (not in default namespace)
-		It("Should fail to read metrics", func() {
-			By("Creating experiment")
-			testName = "invalid-metric"
-			experiment := v2alpha2.NewExperiment(testName, testNamespace).
-				WithTarget("target").
-				WithTestingPattern(v2alpha2.TestingPatternCanary).
-				WithRequestCount("request-count").
-				Build()
-			Expect(k8sClient.Create(ctx(), experiment)).Should(Succeed())
-			By("Checking that it fails")
-			// this depends on an experiment that should run for a while
-			Eventually(func() bool {
-				return containsSubString(events, v2alpha2.ReasonMetricUnavailable) &&
-					containsSubString(events, "default/request-count")
-			}, 5).Should(BeTrue())
-			Eventually(func() bool { return fails(testName, testNamespace) }, 5).Should(BeTrue())
-		})
-	})
-	Context("When creating another experiment which refers to a non-existing metric", func() {
-		// experiment (in default ns) refers to metric "iter8/request-count" (not in iter8 namespace)
-		It("Should fail to read metrics", func() {
-			By("Creating experiment")
-			testName = "invalid-metric"
-			experiment := v2alpha2.NewExperiment(testName, testNamespace).
-				WithTarget("target").
-				WithTestingPattern(v2alpha2.TestingPatternCanary).
-				WithRequestCount("iter8/request-count").
-				Build()
-			Expect(k8sClient.Create(ctx(), experiment)).Should(Succeed())
-			By("Checking that it fails")
-			// this depends on an experiment that should run for a while
-			Eventually(func() bool {
-				return containsSubString(events, v2alpha2.ReasonMetricUnavailable)
-			}, 5).Should(BeTrue())
-			Eventually(func() bool { return fails(testName, testNamespace) }, 5).Should(BeTrue())
-		})
-	})
-
-	Context("When creating an experiment referencing a metric with a bad reference", func() {
-		// experiment (in default namespace) refers to metric "objective-with-bad-reference"
-		// which has a sampleSize "request-count" (not in same ns as the referring metric (default))
-		It("Should fail to read metrics", func() {
-			By("Creating experiment")
-			testName = "invalid-reference"
-			experiment := v2alpha2.NewExperiment(testName, testNamespace).
-				WithTarget("target").
-				WithTestingPattern(v2alpha2.TestingPatternCanary).
-				WithRequestCount(metricsNamespace+"/request-count").
-				WithObjective(*badObjective, nil, nil, false).
-				Build()
-			Expect(k8sClient.Create(ctx(), experiment)).Should(Succeed())
-			By("Checking that it fails")
-			// this depends on an experiment that should run for a while
-			Eventually(func() bool {
-				return containsSubString(events, v2alpha2.ReasonMetricUnavailable)
-			}, 5).Should(BeTrue())
-			// Eventually(func() bool { return fails(testName, testNamespace) }, 5).Should(BeTrue())
-		})
-	})
-
-	Context("When creating an experiment referencing a metric with a bad reference", func() {
-		// experiment (in default namespace) refers to metric "objective-with-bad-reference"
-		// which has a sampleSize "request-count" (not in same ns as the referring metric (default))
-		It("Should successfully read metrics", func() {
-			By("Creating experiment")
-			testName = "good-reference-2"
-
-			experiment := v2alpha2.NewExperiment(testName, testNamespace).
-				WithTarget("target").
-				WithTestingPattern(v2alpha2.TestingPatternCanary).
-				WithRequestCount(metricsNamespace+"/objective-with-good-reference-2").
-				WithObjective(*goodObjective2, nil, nil, false).
-				Build()
-			Expect(k8sClient.Create(ctx(), experiment)).Should(Succeed())
-			By("Checking that it starts Running")
-			// this assumes that it runs for a while
-			Eventually(func() bool {
-				return containsSubString(events, v2alpha2.ReasonStageAdvanced)
-			}, 5).Should(BeTrue())
-		})
-	})
-	Context("When converting a namespacedname", func() {
-		var ns *string
-		var nm string
-		It("Should return nil, '' on input ''", func() {
-			ns, nm = namespaceName("")
-			Expect(ns).To(BeNil())
-			Expect(nm).To(Equal(""))
-		})
-		It("Should return nil, 'name' on input 'name'", func() {
-			ns, nm = namespaceName("name")
-			Expect(ns).To(BeNil())
-			Expect(nm).To(Equal("name"))
-		})
-		It("Should return 'namespace', 'name' on input 'namespace/name", func() {
-			ns, nm = namespaceName("namespace/name")
-			Expect(ns).ToNot(BeNil())
-			Expect(*ns).To(Equal("namespace"))
-			Expect(nm).To(Equal("name"))
-		})
-	})
-})
-
 var _ = Describe("Experiment proceeds", func() {
 	ctx := context.Background()
